pyproject.toml
[project]
name = "price-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "mercadopago>=2.3.0",
    "playwright>=1.58.0",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
]
